[Modify] 01/23 hanjeonghyun 메인 화면 페이지 코드 위치 수정 #7
