+++ app/(anon)/page.tsx
@@ -1,0 +1,18 @@
+"use client";
+
+import React from "react";
+import SubPage from "@/components/index/pages/subPage";
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const ScrollContainer = () => {
+  return (
+    <>
+      <SubPage />
+    </>
+  );
+};
+
+export default ScrollContainer;
