--- app/page.tsx
@@ -1,18 +1,0 @@
-"use client";
-
-import React from "react";
-import SubPage from "@/components/index/pages/subPage";
-
-type Props = {
-  children: React.ReactNode;
-};
-
-const ScrollContainer = () => {
-  return (
-    <>
-      <SubPage />
-    </>
-  );
-};
-
-export default ScrollContainer;
